[pagai] Move fetch_mapping to a common folder
