+++ django/common/mapping/fetch_mapping.py
@@ -5,7 +5,7 @@ from django.conf import settings
 import requests
 from common.errors import OperationOutcome
 
-resource_from_id_query = """
+fragments = """
 fragment entireColumn on Column {
     owner {
         name
@@ -77,7 +77,9 @@ fragment cred on Credential {
     login
     password: decryptedPassword
 }
+"""
 
+resource_from_id_query = """
 query resource($resourceId: String!) {
     resource(where: {id: $resourceId}) {
         id
@@ -108,13 +110,30 @@ query resource($resourceId: String!) {
     }
 }"""
 
+resource_with_filters_query = """
+query resource($resourceId: String!) {
+    resource(where: {id: $resourceId}) {
+        id
+        filters {
+            ...entireFilter
+        }
+        source {
+            id
+            credential {
+                ...cred
+            }
+        }
+    }
+}
+"""
 
-def fetch_resource_mapping(resource_id: str, authorization_header: str):
+
+def run_graphql_query(graphql_query: str, resource_id: str, authorization_header: str):
     try:
         response = requests.post(
             settings.PYROG_API_URL,
             headers={"Authorization": authorization_header},
-            json={"query": resource_from_id_query, "variables": {"resourceId": resource_id}},
+            json={"query": graphql_query, "variables": {"resourceId": resource_id}},
         )
     except requests.exceptions.ConnectionError:
         raise OperationOutcome("Could not connect to the Pyrog service")
@@ -131,9 +150,23 @@ def fetch_resource_mapping(resource_id: str, authorization_header: str):
         else:
             raise OperationOutcome(f"error while fetching mapping: {body['errors']}")
 
-    resource = body["data"]["resource"]
+    return body
+
+
+def fetch_resource_mapping(resource_id: str, authorization_header: str):
+    response = run_graphql_query(f"{fragments}\n{resource_from_id_query}", resource_id, authorization_header)
+    resource = response["data"]["resource"]
 
     dereference_concept_map(resource, authorization_header)
+
+    if not resource:
+        raise OperationOutcome(f"resource with id {resource_id} does not exist")
+    return resource
+
+
+def fetch_resource_with_filters(resource_id: str, authorization_header: str):
+    response = run_graphql_query(f"{fragments}\n{resource_with_filters_query}", resource_id, authorization_header)
+    resource = response["data"]["resource"]
 
     if not resource:
         raise OperationOutcome(f"resource with id {resource_id} does not exist")
