+++ tests/common/test_fetch_mapping.py
@@ -1,19 +1,19 @@
 from pathlib import Path
 
+from pytest import raises
 from rest_framework.exceptions import NotAuthenticated, PermissionDenied
 
-from control.api import fetch_mapping
-from pytest import raises
+from common.mapping import fetch_mapping
 
 FIXTURES_DIR = Path(__file__).resolve().parent / "fixtures"
 
 
-def test_fetch_resource_mapping_invalid(mock_fhir_api_response):
+def test_fetch_resource_mapping_invalid(mock_pyrog_response, mock_fhir_api_response):
     with raises(NotAuthenticated):
         fetch_mapping.fetch_concept_map("id", "Bearer invalidToken")
 
 
-def test_fetch_resource_mapping_forbidden(mock_fhir_api_response):
+def test_fetch_resource_mapping_forbidden(mock_pyrog_response, mock_fhir_api_response):
     with raises(PermissionDenied):
         fetch_mapping.fetch_concept_map("id", "Bearer forbiddenToken")
